fix out of range err

+++ migrations/000001_create_metrics_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE metric(
     name varchar(255),
     "type" varchar(7), -- Можно потом сделать оптимальнее
     "value" double precision,
-    delta integer,
+    delta BIGINT,
     PRIMARY KEY(id)
 );
 CREATE UNIQUE INDEX type_name_1762799044404_index ON public.metric USING btree (type, name);
